+++ docarray/util/__init__.py
@@ -1,0 +1,3 @@
+from docarray.util.find import find, find_batched
+
+__all__ = ['find', 'find_batched']
