--- docarray/utility/__init__.py
@@ -1,3 +1,0 @@
-from docarray.utility.find import find, find_batched
-
-__all__ = ['find', 'find_batched']
