--- docarray/utility/distance.py
@@ -1,1 +1,0 @@
-
